Refactor: Fix casting to the same type is unnecessary

+++ src-tauri/src/commands/hardware.rs
@@ -94,7 +94,7 @@ pub fn get_process_list(state: tauri::State<'_, AppState>) -> Vec<ProcessInfo> {
       let memory_usage = if let Some(history) = process_memory_histories.get(&pid) {
         let len = history.len().min(5); // 最大5秒分のデータ
         let sum: f32 = history.iter().rev().take(len).sum();
-        let avg = (sum as f32 / 1024.0) / len as f32;
+        let avg = (sum / 1024.0) / len as f32;
 
         (avg * 10.0).round() / 10.0 // 小数点1桁で丸める
       } else {
@@ -205,7 +205,7 @@ pub fn get_memory_usage(state: tauri::State<'_, AppState>) -> i32 {
   let used_memory = system.used_memory() as f64;
   let total_memory = system.total_memory() as f64;
 
-  ((used_memory / total_memory) * 100.0 as f64).round() as i32
+  ((used_memory / total_memory) * 100.0).round() as i32
 }
 
 ///
@@ -380,7 +380,7 @@ pub fn initialize_system(
       let cpu_usage = {
         let cpus = sys.cpus();
         let total_usage: f32 = cpus.iter().map(|cpu| cpu.cpu_usage()).sum();
-        (total_usage / cpus.len() as f32).round() as f32
+        (total_usage / cpus.len() as f32).round()
       };
 
       let memory_usage = {
@@ -412,7 +412,7 @@ pub fn initialize_system(
 
         for (pid, process) in sys.processes() {
           // CPU使用率の履歴を更新
-          let cpu_usage = process.cpu_usage() as f32;
+          let cpu_usage = process.cpu_usage();
           let cpu_history = process_cpu_histories.entry(*pid).or_default();
 
           if cpu_history.len() >= HISTORY_CAPACITY {
